Rename `inner.rs` as `calls.rs`
Because it's a clearer name.

Also move some call-related stuff from `plumbing.rs` to `calls.rs`.
