+++ compiler/rustc_middle/src/query/mod.rs
@@ -1,21 +1,21 @@
 use rustc_hir::def_id::LocalDefId;
 
 pub use self::caches::{DefIdCache, DefaultCache, QueryCache, SingleCache, VecCache};
+pub use self::calls::{TyCtxtAt, TyCtxtEnsureDone, TyCtxtEnsureOk, TyCtxtEnsureResult};
 pub use self::into_query_key::IntoQueryKey;
 pub use self::job::{
     ActiveKeyStatus, QueryCycle, QueryJob, QueryJobId, QueryLatch, QueryStackFrame, QueryState,
     QueryWaiter,
 };
 pub use self::keys::{LocalCrate, QueryKey};
-pub use self::plumbing::{TyCtxtAt, TyCtxtEnsureDone, TyCtxtEnsureOk, TyCtxtEnsureResult};
 pub use self::system::{QueryMode, QuerySystem, QueryVTable};
 pub use crate::queries::Providers;
 use crate::ty::TyCtxt;
 
 pub(crate) mod arena_cached;
 mod caches;
+pub(crate) mod calls;
 pub mod erase;
-pub(crate) mod inner;
 mod into_query_key;
 mod job;
 mod keys;
