+++ compiler/rustc_middle/src/query/plumbing.rs
@@ -1,117 +1,3 @@
-use std::ops::Deref;
-
-use rustc_hir::def_id::LocalDefId;
-use rustc_span::Span;
-
-use crate::query::IntoQueryKey;
-use crate::ty::{self, TyCtxt};
-
-#[derive(Copy, Clone)]
-pub struct TyCtxtAt<'tcx> {
-    pub tcx: TyCtxt<'tcx>,
-    pub span: Span,
-}
-
-impl<'tcx> Deref for TyCtxtAt<'tcx> {
-    type Target = TyCtxt<'tcx>;
-    #[inline(always)]
-    fn deref(&self) -> &Self::Target {
-        &self.tcx
-    }
-}
-
-#[derive(Copy, Clone)]
-#[must_use]
-pub struct TyCtxtEnsureOk<'tcx> {
-    pub tcx: TyCtxt<'tcx>,
-}
-
-#[derive(Copy, Clone)]
-#[must_use]
-pub struct TyCtxtEnsureResult<'tcx> {
-    pub tcx: TyCtxt<'tcx>,
-}
-
-#[derive(Copy, Clone)]
-#[must_use]
-pub struct TyCtxtEnsureDone<'tcx> {
-    pub tcx: TyCtxt<'tcx>,
-}
-
-impl<'tcx> TyCtxtEnsureOk<'tcx> {
-    pub fn typeck(self, def_id: impl IntoQueryKey<LocalDefId>) {
-        self.typeck_root(
-            self.tcx.typeck_root_def_id(def_id.into_query_key().to_def_id()).expect_local(),
-        )
-    }
-}
-
-impl<'tcx> TyCtxt<'tcx> {
-    pub fn typeck(self, def_id: impl IntoQueryKey<LocalDefId>) -> &'tcx ty::TypeckResults<'tcx> {
-        self.typeck_root(
-            self.typeck_root_def_id(def_id.into_query_key().to_def_id()).expect_local(),
-        )
-    }
-
-    /// Returns a transparent wrapper for `TyCtxt` which uses
-    /// `span` as the location of queries performed through it.
-    #[inline(always)]
-    pub fn at(self, span: Span) -> TyCtxtAt<'tcx> {
-        TyCtxtAt { tcx: self, span }
-    }
-
-    /// FIXME: `ensure_ok`'s effects are subtle. Is this comment fully accurate?
-    ///
-    /// Wrapper that calls queries in a special "ensure OK" mode, for callers
-    /// that don't need the return value and just want to invoke a query for
-    /// its potential side-effect of emitting fatal errors.
-    ///
-    /// This can be more efficient than a normal query call, because if the
-    /// query's inputs are all green, the call can return immediately without
-    /// needing to obtain a value (by decoding one from disk or by executing
-    /// the query).
-    ///
-    /// (As with all query calls, execution is also skipped if the query result
-    /// is already cached in memory.)
-    ///
-    /// ## WARNING
-    /// A subsequent normal call to the same query might still cause it to be
-    /// executed! This can occur when the inputs are all green, but the query's
-    /// result is not cached on disk, so the query must be executed to obtain a
-    /// return value.
-    ///
-    /// Therefore, this call mode is not appropriate for callers that want to
-    /// ensure that the query is _never_ executed in the future.
-    #[inline(always)]
-    pub fn ensure_ok(self) -> TyCtxtEnsureOk<'tcx> {
-        TyCtxtEnsureOk { tcx: self }
-    }
-
-    /// This is a variant of `ensure_ok` only usable with queries that return
-    /// `Result<_, ErrorGuaranteed>`. Queries calls through this function will
-    /// return `Result<(), ErrorGuaranteed>`. I.e. the error status is returned
-    /// but nothing else. As with `ensure_ok`, this can be more efficient than
-    /// a normal query call.
-    #[inline(always)]
-    pub fn ensure_result(self) -> TyCtxtEnsureResult<'tcx> {
-        TyCtxtEnsureResult { tcx: self }
-    }
-
-    /// Wrapper that calls queries where callers don't need the return value and
-    /// just want to guarantee that the query won't be executed in the future.
-    ///
-    /// This is useful for queries that read from a [`Steal`] value, to ensure
-    /// that they are executed before the query that will steal the value.
-    ///
-    /// Currently this causes the query to be executed normally, but this behavior may change.
-    ///
-    /// [`Steal`]: rustc_data_structures::steal::Steal
-    #[inline(always)]
-    pub fn ensure_done(self) -> TyCtxtEnsureDone<'tcx> {
-        TyCtxtEnsureDone { tcx: self }
-    }
-}
-
 macro_rules! maybe_into_query_key {
     (DefId) => { impl $crate::query::IntoQueryKey<DefId> };
     (LocalDefId) => { impl $crate::query::IntoQueryKey<LocalDefId> };
@@ -398,9 +284,7 @@ macro_rules! define_callbacks {
                 $(#[$attr])*
                 #[inline(always)]
                 pub fn $name(self, key: maybe_into_query_key!($($K)*)) -> $V {
-                    use $crate::query::{erase, inner};
-
-                    erase::restore_val::<$V>(inner::query_get_at(
+                    $crate::query::erase::restore_val::<$V>($crate::query::calls::query_get_at(
                         self.tcx,
                         self.span,
                         &self.tcx.query_system.query_vtables.$name,
@@ -415,7 +299,7 @@ macro_rules! define_callbacks {
                 $(#[$attr])*
                 #[inline(always)]
                 pub fn $name(self, key: maybe_into_query_key!($($K)*)) {
-                    $crate::query::inner::query_ensure_ok(
+                    $crate::query::calls::query_ensure_ok(
                         self.tcx,
                         &self.tcx.query_system.query_vtables.$name,
                         $crate::query::IntoQueryKey::into_query_key(key),
@@ -434,7 +318,7 @@ macro_rules! define_callbacks {
                     self,
                     key: maybe_into_query_key!($($K)*),
                 ) -> Result<(), rustc_errors::ErrorGuaranteed> {
-                    $crate::query::inner::query_ensure_result(
+                    $crate::query::calls::query_ensure_result(
                         self.tcx,
                         &self.tcx.query_system.query_vtables.$name,
                         $crate::query::IntoQueryKey::into_query_key(key),
@@ -464,7 +348,7 @@ macro_rules! define_callbacks {
                 $(#[$attr])*
                 #[inline(always)]
                 pub fn $name(self, value: $name::ProvidedValue<'tcx>) {
-                    $crate::query::inner::query_feed(
+                    $crate::query::calls::query_feed(
                         self.tcx,
                         &self.tcx.query_system.query_vtables.$name,
                         self.key().into_query_key(),
